+++ src/EFCore/Query/Internal/QueryOptimizingExpressionVisitor.cs
@@ -9,15 +9,50 @@ using Microsoft.EntityFrameworkCore.Utilities;
 
 namespace Microsoft.EntityFrameworkCore.Query.Internal
 {
-    public class AllAnyContainsRewritingExpressionVisitor : ExpressionVisitor
+    public class QueryOptimizingExpressionVisitor : ExpressionVisitor
     {
-        private static bool IsExpressionOfFunc(Type type, int funcGenericArgs = 2)
-            => type.IsGenericType
-                && type.GetGenericArguments().Length == funcGenericArgs;
+        private static readonly MethodInfo _stringCompareWithComparisonMethod =
+            typeof(string).GetRuntimeMethod(nameof(string.Compare), new[] { typeof(string), typeof(string), typeof(StringComparison) });
+        private static readonly MethodInfo _stringCompareWithoutComparisonMethod =
+            typeof(string).GetRuntimeMethod(nameof(string.Compare), new[] { typeof(string), typeof(string) });
+        private static readonly MethodInfo _startsWithMethodInfo =
+            typeof(string).GetRuntimeMethod(nameof(string.StartsWith), new[] { typeof(string) });
+        private static readonly MethodInfo _endsWithMethodInfo =
+            typeof(string).GetRuntimeMethod(nameof(string.EndsWith), new[] { typeof(string) });
+
+        private static readonly Expression _constantNullString = Expression.Constant(null, typeof(string));
 
         protected override Expression VisitMethodCall(MethodCallExpression methodCallExpression)
         {
             Check.NotNull(methodCallExpression, nameof(methodCallExpression));
+
+            if (_startsWithMethodInfo.Equals(methodCallExpression.Method)
+                || _endsWithMethodInfo.Equals(methodCallExpression.Method))
+            {
+                if (methodCallExpression.Arguments[0] is ConstantExpression constantArgument
+                    && (string)constantArgument.Value == string.Empty)
+                {
+                    // every string starts/ends with empty string.
+                    return Expression.Constant(true);
+                }
+
+                var newObject = Visit(methodCallExpression.Object);
+                var newArgument = Visit(methodCallExpression.Arguments[0]);
+
+                var result = Expression.AndAlso(
+                    Expression.NotEqual(newObject, _constantNullString),
+                    Expression.AndAlso(
+                        Expression.NotEqual(newArgument, _constantNullString),
+                        methodCallExpression.Update(newObject, new[] { newArgument })));
+
+                return newArgument is ConstantExpression
+                    ? result
+                    : Expression.OrElse(
+                        Expression.Equal(
+                            newArgument,
+                            Expression.Constant(string.Empty)),
+                        result);
+            }
 
             if (methodCallExpression.Method.IsGenericMethod
                 && methodCallExpression.Method.GetGenericMethodDefinition() is MethodInfo methodInfo
@@ -46,9 +81,7 @@ namespace Microsoft.EntityFrameworkCore.Query.Internal
 
             if (methodCallExpression.Method.IsGenericMethod
                 && methodCallExpression.Method.GetGenericMethodDefinition() is MethodInfo containsMethodInfo
-                && containsMethodInfo.Equals(QueryableMethods.Contains)
-                // special case Queryable.Contains(byte_array, byte) - we don't want those to be rewritten
-                && methodCallExpression.Arguments[1].Type != typeof(byte))
+                && containsMethodInfo.Equals(QueryableMethods.Contains))
             {
                 var typeArgument = methodCallExpression.Method.GetGenericArguments()[0];
                 var anyMethod = QueryableMethods.AnyWithPredicate.MakeGenericMethod(typeArgument);
@@ -63,7 +96,67 @@ namespace Microsoft.EntityFrameworkCore.Query.Internal
                 return Expression.Call(null, anyMethod, new[] { methodCallExpression.Arguments[0], anyLambda });
             }
 
-            return base.VisitMethodCall(methodCallExpression);
+            var visited = (MethodCallExpression)base.VisitMethodCall(methodCallExpression);
+
+            // In VB.NET, comparison operators between strings (equality, greater-than, less-than) yield
+            // calls to a VB-specific CompareString method. Normalize that to string.Compare.
+            if (visited.Method.Name == "CompareString"
+                && visited.Method.DeclaringType?.Name == "Operators"
+                && visited.Method.DeclaringType?.Namespace == "Microsoft.VisualBasic.CompilerServices"
+                && visited.Object == null
+                && visited.Arguments.Count == 3
+                && visited.Arguments[2] is ConstantExpression textCompareConstantExpression)
+            {
+                return (bool)textCompareConstantExpression.Value
+                    ? Expression.Call(
+                        _stringCompareWithComparisonMethod,
+                        visited.Arguments[0],
+                        visited.Arguments[1],
+                        Expression.Constant(StringComparison.OrdinalIgnoreCase))
+                    : Expression.Call(
+                        _stringCompareWithoutComparisonMethod,
+                        visited.Arguments[0],
+                        visited.Arguments[1]);
+            }
+
+            return visited;
+        }
+
+        protected override Expression VisitUnary(UnaryExpression unaryExpression)
+        {
+            Check.NotNull(unaryExpression, nameof(unaryExpression));
+
+            if (unaryExpression.NodeType == ExpressionType.Not
+                && unaryExpression.Operand is MethodCallExpression innerMethodCall
+                && (_startsWithMethodInfo.Equals(innerMethodCall.Method)
+                    || _endsWithMethodInfo.Equals(innerMethodCall.Method)))
+            {
+                if (innerMethodCall.Arguments[0] is ConstantExpression constantArgument
+                    && (string)constantArgument.Value == string.Empty)
+                {
+                    // every string starts/ends with empty string.
+                    return Expression.Constant(false);
+                }
+
+                var newObject = Visit(innerMethodCall.Object);
+                var newArgument = Visit(innerMethodCall.Arguments[0]);
+
+                var result = Expression.AndAlso(
+                    Expression.NotEqual(newObject, _constantNullString),
+                    Expression.AndAlso(
+                        Expression.NotEqual(newArgument, _constantNullString),
+                        Expression.Not(innerMethodCall.Update(newObject, new[] { newArgument }))));
+
+                return newArgument is ConstantExpression
+                    ? result
+                    : Expression.AndAlso(
+                        Expression.NotEqual(
+                            newArgument,
+                            Expression.Constant(string.Empty)),
+                        result);
+            }
+
+            return base.VisitUnary(unaryExpression);
         }
 
         private bool TryExtractEqualityOperands(Expression expression, out Expression left, out Expression right, out bool negated)
